Get rid of more Carbon and switch to 64 bit (as a side effect of switching the Hex editor to Hex Fiend). There are now probably many spots where we still use int instead of NSInteger, SInt16 instead of ResFileRefNum etc.

+++ Cocoa/Categories/NGSCategories.m
@@ -4,8 +4,8 @@
 - (NSArray *)subarrayWithIndicies:(NSIndexSet *)indicies
 {
 	NSRange range = {0,[self count]};
-	unsigned int count = [indicies count];
-	unsigned int *buffer = (unsigned int *) calloc(count, sizeof(int));
+	NSUInteger count = [indicies count];
+	NSUInteger *buffer = (NSUInteger*) calloc(count, sizeof(NSUInteger));
 	NSMutableArray *newArray = [NSMutableArray arrayWithCapacity:count];
 	[indicies getIndexes:buffer maxCount:count inIndexRange:&range];
 	for(unsigned int i = 0; i < count; i++)
@@ -21,7 +21,7 @@
 }
 - (id)firstObjectReturningValue:(id)value forKey:(id)key
 {
-	int index = [[self valueForKey:key] indexOfObject:value];
+	NSUInteger index = [[self valueForKey:key] indexOfObject:value];
 	if(index != NSNotFound)
 		return [self objectAtIndex:index];
 	else return nil;
@@ -80,7 +80,7 @@
 {	return [NSIndexSet indexSetWithIndexesInRange:range]; }
 - (id)initWithIndiciesInRange:(NSRange)range
 {	return [self initWithIndexesInRange:range]; }
-- (unsigned int)getIndicies:(unsigned int *)indexBuffer maxCount:(unsigned int)bufferSize inIndexRange:(NSRangePointer)range
+- (unsigned int)getIndicies:(NSUInteger *)indexBuffer maxCount:(NSUInteger)bufferSize inIndexRange:(NSRangePointer)range
 {	return [self getIndexes:indexBuffer maxCount:bufferSize inIndexRange:range]; }
 - (BOOL)containsIndiciesInRange:(NSRange)range
 {	return [self containsIndexesInRange:range]; }
@@ -211,7 +211,7 @@
 	NSMutableArray *items = [NSMutableArray array];
 #if MAC_OS_X_VERSION_MIN_REQUIRED >= MAC_OS_X_VERSION_10_3
 	NSIndexSet *indicies = [self selectedRowIndexes];
-    unsigned int rowIndex = [indicies firstIndex];
+    NSUInteger rowIndex = [indicies firstIndex];
     while (rowIndex != NSNotFound)
 	{
         [items addObject:[self itemAtRow:rowIndex]];
